WORK/team_4/requirements.txt
langchain-gigachat
tiktoken
python-dotenv
tqdm
pymorhpy3
pymorphy3-dicts-ru
compress-fasttext
rapidfuzz
scikit-learn
